+++ src/Base/isBrowser/isBrowser.ts
@@ -1,0 +1,9 @@
+/**
+ * 判断是否是浏览器
+ * @returns
+ * @example
+ *
+ */
+export default function isBrowser(): boolean {
+	return typeof window === 'object' && typeof document === 'object';
+}