+++ src/Math/radsToDegs/radsToDegs.ts
@@ -1,0 +1,9 @@
+/**
+ * 将弧度转换为角度
+ * @returns
+ * @example
+ *
+ */
+export default function radsToDegs(rad: number): number {
+	return (rad * 180) / Math.PI;
+}